refactor: extract block_device
